fix(desktop): address Quick Pick PR review feedback
Merge result UI into a single quick-pick window, load locale via invoke
for lightweight windows, cache Enigo for cursor position, and harden
focus handling, insert paste, error loading, and cancel UX.
