+++ apps/desktop/src-tauri/capabilities/desktop.json
@@ -9,8 +9,7 @@
     "main",
     "indicator",
     "upgrade",
-    "quick-pick",
-    "quick-pick-result"
+    "quick-pick"
   ],
   "permissions": [
     "global-shortcut:default",
