+++ apps/desktop/src/App.vue
@@ -11,7 +11,6 @@ import { syncTrayMenu } from '@/tray'
 import Indicator from '@/windows/Indicator.vue'
 import Main from '@/windows/Main.vue'
 import QuickPick from '@/windows/QuickPick.vue'
-import QuickPickResult from '@/windows/QuickPickResult.vue'
 import Upgrade from '@/windows/Upgrade.vue'
 
 const appWindow = getCurrentWebviewWindow()
@@ -22,14 +21,13 @@ const windows: Record<string, Component> = {
   'indicator': Indicator,
   'upgrade': Upgrade,
   'quick-pick': QuickPick,
-  'quick-pick-result': QuickPickResult,
 }
 
 let isMounted = true
 onMounted(async () => {
   logger.info('App', `onMounted for window: ${currentLabel}`)
 
-  const isQuickPickWindow = currentLabel === 'quick-pick' || currentLabel === 'quick-pick-result'
+  const isQuickPickWindow = currentLabel === 'quick-pick'
 
   // Quick-pick windows skip store/auth (no server sync); locale via Rust settings read.
   if (!isQuickPickWindow) {
